Add `Sai::Decorator::NamedColors`
