+++ sig/sai/decorator.rbs
@@ -8,6 +8,8 @@ module Sai
   #
   # @api public
   class Decorator
+    include NamedColors
+
     # Initialize a new instance of Decorator
     #
     # @author {https://aaronmallen.me Aaron Allen}
@@ -20,70 +22,6 @@ module Sai
     # @return [Decorator] the new instance of Decorator
     # @rbs (?mode: Integer) -> void
     def initialize: (?mode: Integer) -> void
-
-    def black: () -> Decorator
-
-    def blue: () -> Decorator
-
-    def bright_black: () -> Decorator
-
-    def bright_blue: () -> Decorator
-
-    def bright_cyan: () -> Decorator
-
-    def bright_green: () -> Decorator
-
-    def bright_magenta: () -> Decorator
-
-    def bright_red: () -> Decorator
-
-    def bright_white: () -> Decorator
-
-    def bright_yellow: () -> Decorator
-
-    def cyan: () -> Decorator
-
-    def green: () -> Decorator
-
-    def magenta: () -> Decorator
-
-    def on_black: () -> Decorator
-
-    def on_blue: () -> Decorator
-
-    def on_bright_black: () -> Decorator
-
-    def on_bright_blue: () -> Decorator
-
-    def on_bright_cyan: () -> Decorator
-
-    def on_bright_green: () -> Decorator
-
-    def on_bright_magenta: () -> Decorator
-
-    def on_bright_red: () -> Decorator
-
-    def on_bright_white: () -> Decorator
-
-    def on_bright_yellow: () -> Decorator
-
-    def on_cyan: () -> Decorator
-
-    def on_green: () -> Decorator
-
-    def on_magenta: () -> Decorator
-
-    def on_red: () -> Decorator
-
-    def on_white: () -> Decorator
-
-    def on_yellow: () -> Decorator
-
-    def red: () -> Decorator
-
-    def white: () -> Decorator
-
-    def yellow: () -> Decorator
 
     def blink: () -> Decorator
 
@@ -407,20 +345,6 @@ module Sai
     # @return [Array<Array<Integer>>] adjusted color sequence
     # @rbs (Array[Array[Integer]] colors, Integer text_length) -> Array[Array[Integer]]
     def adjust_colors_to_text_length: (Array[Array[Integer]] colors, Integer text_length) -> Array[Array[Integer]]
-
-    # Apply a named color to the specified style type
-    #
-    # @author {https://aaronmallen.me Aaron Allen}
-    # @since 0.1.0
-    #
-    # @api private
-    #
-    # @param style_type [Symbol] the style type to apply the color to
-    # @param color [Symbol] the color to apply
-    #
-    # @return [Decorator] a new instance of Decorator with the color applied
-    # @rbs (Conversion::ColorSequence::style_type style_type, Symbol color) -> Decorator
-    def apply_named_color: (Conversion::ColorSequence::style_type style_type, Symbol color) -> Decorator
 
     # Apply color sequence gradients to text
     #
